refactor: address PR review feedback on SSR feature flags
- Centralize defaultFeatures in api-client.ts; import it in both the
  client context and the server helper (removes duplication)
- features-server.ts: strip trailing slash from the API base URL, lower
  the SSR fetch timeout 2s -> 1s, guard against non-object JSON payloads
  before spreading, and log fetch failures instead of swallowing them
- layout.tsx: resolve getLocale/getMessages/getFeaturesServerSide with
  Promise.all so the flag fetch adds no latency beyond what the layout
  already awaits (mitigates blocking every route on a cold backend)
